more work with DTOs and DI

+++ app/src/main/java/com/example/mobileup_trainee_test_task/domain/models/CryptoCurrency.kt
@@ -6,7 +6,7 @@ data class CryptoCurrency(
     val symbol: String,
     val name: String,
     val image: String,
-    val currentPrice: Int,
+    val currentPrice: Double,
     val priceChange24h: Double,
     ) {
 }
